Change feedback form links
This could be an environment variable, but given it was hard-coded in
the repo we forked from, it might be it doesn't need to change often
enough to justify it.

Given the other things we have to address, I'd rather change this to
an env-var if it comes up again soon.