Add static I-need-help URL for integration tests (#13)
* Add static I-need-help URL for integration tests

The integration tests currently rely on a INH_URL environment
variable being a particular value. Since this variable is used to
configure the real link that users see, I changed it to reflect our
real environment - and the integration tests started failing.

Unfortunately, the configuration in setupTests.js don't seem to
affect the integration tests.

This commit adds a hard-coded URL to the env vars passed to the
integration tests, matching their expectations. I suspect there
may be a better long-term way to do this.

* add inh url for staging and production

* amend inh url

* amend test with inh url env var

Co-authored-by: Riccardo Noviello <riccardo.noviello@gmail.com>