Only show TopicExplorer if feature flag is set
This commit is experimental, based on my meagre understanding of
NextJS's interaction with environment variables.

From the docs (https://nextjs.org/docs/basic-features/environment-
variables#exposing-environment-variables-to-the-browser):
"In order to expose a variable to the browser you have to prefix the
variable with NEXT_PUBLIC_."

This works locally, but I believe our pipeline may pre-compile the
client-side version of the app - in this case even changing the env
var in the lambda that runs this app may not change the behaviour.

My intent is to push this to an AWS environment and then see what
happens!