Load from env var in Index's getInitialProps
This commit adds another prop to the Index component, attempting to
make showing the TopicExplorer component consistent between the
server- and client-side versions of the app.

It also changes the env var to cut of the NEXT_PUBLIC_ prefix, since
this implementation no longer relies on nextjs magic to expose env
vars to a client-side app.