Retrieve prompts data from Airtable (#23)
* Retrieve prompts data from airtable

* Get prompts from airtable

This commit is aimed at abstracting prompts from resources. The gateway
containes an `all` function which gets all the data from all tables
with a given `AIRTABLE_BASE_ID`. This commit adds another function to
the gateway which allows you to request data from a given table alowing
for clear abstraction between the two concepts.

* Abstract prompt domain object from resource domain

* Remove concept of prompts from componets

This will be reintroduced as the next piece of work.

* Add empty line at EOF and format

* Add env var to env sample