Without search term, show prompts tagged 'default'
Tripped up here for a while around useEffect(). Without the argument
on TopicExplorer/index.js:42, this hook runs infinitely, since setting
the new state causes the component to re-render (and thereby run
useEffect() again).

This added argument says "only run this if one of these states
changed", and since only search RESULTS changed, not the term, it
works as intended.

The nearby implementation of tagging dodges this because of the if
statement on line 35 - I tried to re-write this as an accumulator,
and it also looped!