Use hard-coded path rather than env var
NEXT_PUBLIC_URL is returning undefined for a reason that's eluding me
- I suspect it's that this env var isn't available client-side, but I
haven't been able to resolve it in the time available.

This commit risks someone being directed to a live environment
accidentally upon logging in, but I think it is worth the tradeoff for
users not getting stuck/confused after they login.