Add ref number to emails and receipts (#129)
* Added dynamodb spin-up script to package.json

* Added Reference Number to Call Handler's Email Receipt.

To achieve this, the generation of reference number had to moved from the 'referral-gateway' to the 'create-referral' use-case. It's because we want that id to be passed into multiple code execution branches branching out from this use-case.

* Added reference number to Referal Email Template.

* Added reference number into a Success confirmation message & Accepted/Rejected referal pages.

* Rename referenceCode to referenceNumber

* Fixed failing gateway & usecase unit tests for create-referal.

They were failing because the referenceNumber argument was unaccounted for.

* Fixed cypress integration tests for accept/reject referal functionality.

They were failing because the static text element that changed was tested. It ended up working in our favour, because we replaced that text with text containing important variable. Also had to add that variable (referenceNumber) into mock data that gets fed into local dynamodb.

Co-authored-by: Kat <katrina@madetech.com>