Added a function that returns a list of words to be highlighted in the correct casing
This function is only the 1st draft. It badly needs refactoring to extract repeated code, also it's not tested. Not sure if it works or not as I didn't have time to test it, hope it's alright. Work mainly came into the idea behind it anyway. Sorry to leave this like this, but I'm taking a sick day >_<