Fixed the address line 2 single referral page test.
In short, when searching for 'Hackney' it has likely found some other node on the VDOM that was causing a crash within the internals of that 'search the document' function. After changing 'Hackney' to 'London' that problem node no longer gets picked up, while we're still testing the same thing just under different text.