Added aria atribute to link reject reason input with its label.
Even though the label has 'for' attribute, the 'for' attribute for reason looks at for a text input with that 'id' rather than 'name'. This somehow started causing accessibility tests to start failing because they didn't see them linked. I suspect it's because I added 'id' attribute, which gets prioritized over 'name'. Problem is that even though 'id' looks identical in code, once it's on the page it has a component or page name as well as hash appended to it by css modules behaviour.