Bypass a bug within Wiremock's 'install.js' script.
If wiremock's version isn't explictly specified, then wiremock attempts to install the latest version available. It gets this information by making a request to get the maven-metadata.xml file. The regex that extracts the latest version from XML expects it to be in the typical 3 dot-separated digit format sandwitched between the 'release' XML tags. However, that is not the case currently. The regex pattern fails due to unexpected '-beta-2' suffix after the 3 dot-separated digits. The script fails as it doesn't know what version to download.